+++ src/simplebuffer.h
@@ -26,11 +26,14 @@
 #define MMAP_NOT_SHMGET
 struct simplebuf{
   struct opt_s *opt;
-  int diff;
+  /* Migrated diffs to work as number of bytes */
+  long diff;
+  //int diff;
   /* Used in init and close. Otherwise opts optbits is used 	*/
   /* TODO: Needs some refactoring to make more sensible		*/
   int optbits;
-  int asyncdiff;
+  long asyncdiff;
+  //int asyncdiff;
   void* buffer;
   int async_writes_submitted;
   int running;
@@ -54,6 +57,6 @@ struct simplebuf{
 };
 int sbuf_init(struct opt_s *opt, struct buffer_entity *be);
 int sbuf_close(struct buffer_entity *be, void * stats);
-void* sbuf_getbuf(struct buffer_entity *be, int** diff);
+void* sbuf_getbuf(struct buffer_entity *be, long** diff);
 int sbuf_init_buf_entity(struct opt_s *opt, struct buffer_entity *be);
 #endif
